refactor(a2a): split adapter impl into bounded modules (wave8a step2)
